Decouple pool from service
